requirements.txt
ray[serve]
scikit-learn
faiss-cpu
numpy